refactor(glm52): hoist derived constants to config.rs, extract build module
ROPE_HALF, SM_SCALE, INDEX_HEAD_DIM, rope_tables, and
glm52_layer_has_full_indexer were duplicated across model/mod.rs,
oracle/{mla,layer,layer_ep8,indexer}.rs, and indexer_smoke.rs — each
copy a hand-maintained literal disconnected from the config-validated
constants they derive from.

- config.rs: add GLM52_ROPE_HALF (= QK_ROPE_HEAD_DIM/2), GLM52_SM_SCALE
  (= 1/sqrt(QK_HEAD_DIM)), move glm52_layer_has_full_indexer here (now
  expressed via INDEX_SKIP_TOPK_OFFSET/INDEX_TOPK_FREQ instead of
  hardcoded 2 and 4), expose GLM52_ROPE_THETA as pub(crate).
- model/mod.rs: drop the 5 duplicate locals; rope_tables is the single
  implementation. INDEX_CACHE_BLOCK (DeepGEMM kernel constraint) and
  NUM_SMS (H200 hardware property) stay here — not model properties.
- Extract build_decoder_layer + take_proj + take_bf16_vec into
  model/build.rs (mod.rs 956 -> 816 lines, build/execute separation).
- oracle/, indexer_smoke.rs: remove local constant copies, import from
  config + model.
