+++ openinfer-glm52/src/config.rs
@@ -24,6 +24,13 @@ pub(crate) const GLM52_KV_A_OUT: usize = GLM52_KV_LORA_RANK + GLM52_QK_ROPE_HEAD
 pub(crate) const GLM52_KV_B_OUT: usize = GLM52_HEADS * (GLM52_QK_NOPE_HEAD_DIM + GLM52_V_HEAD_DIM);
 pub(crate) const GLM52_O_PROJ_IN: usize = GLM52_HEADS * GLM52_V_HEAD_DIM;
 
+/// Half of the rotary head dim — the rotary table width shared by MLA and
+/// indexer rope.
+pub(crate) const GLM52_ROPE_HALF: usize = GLM52_QK_ROPE_HEAD_DIM / 2;
+/// `1/sqrt(GLM52_QK_HEAD_DIM)` = 1/sqrt(256) = 0.0625 — the MLA softmax scale
+/// (rope_type "default" means no YaRN mscale correction).
+pub(crate) const GLM52_SM_SCALE: f32 = 0.0625;
+
 pub(crate) const GLM52_DENSE_INTERMEDIATE: usize = 12_288;
 pub(crate) const GLM52_EXPERT_INTERMEDIATE: usize = 2048;
 pub const GLM52_ROUTED_EXPERTS: usize = 256;
@@ -42,7 +49,16 @@ pub(crate) const GLM52_INDEX_HEADS: usize = 32;
 const GLM52_INDEX_SKIP_TOPK_OFFSET: usize = 3;
 const GLM52_NEXTN_LAYERS: usize = 1;
 
-const GLM52_ROPE_THETA: f64 = 8_000_000.0;
+pub(crate) const GLM52_ROPE_THETA: f64 = 8_000_000.0;
+
+/// `indexer_types[layer]` per the transformers derivation
+/// (`index_topk_freq=4`, `index_skip_topk_offset=3`): full iff
+/// `max(layer-(offset-1), 0) % freq == 0` → {0,1,2} ∪ {6,10,…,74}, 21 of 78 layers.
+pub(crate) fn glm52_layer_has_full_indexer(layer: usize) -> bool {
+    layer
+        .saturating_sub(GLM52_INDEX_SKIP_TOPK_OFFSET - 1)
+        .is_multiple_of(GLM52_INDEX_TOPK_FREQ)
+}
 
 pub fn probe_config_json(json: &Value) -> Result<()> {
     let model_type = string_field(json, "model_type")?;
