+++ openinfer-glm52/src/oracle/indexer.rs
@@ -32,8 +32,10 @@ use sha2::{Digest, Sha256};
 use openinfer_kernels::ops::{GLM52_INDEXER_TOPK, Glm52IndexerCacheLayout};
 use openinfer_kernels::tensor::DeviceContext;
 
+use crate::config::GLM52_INDEX_HEAD_DIM;
 use crate::fp8::Glm52ProjBytes;
 use crate::indexer::{Glm52IndexerLayerWeights, glm52_indexer_cache_fill, glm52_indexer_forward};
+use crate::model::{INDEX_CACHE_BLOCK, NUM_SMS};
 
 // ---- oracle provenance ----
 // Generated by: tools/accuracy/glm52_oracle.py --model-path /data/models/GLM-5.2-FP8 \
@@ -45,9 +47,6 @@ const ORACLE_HIDDEN_DIGEST: &str = "7f1235f1bf66b263";
 
 const HIDDEN: usize = 6144;
 const Q_LORA: usize = 2048;
-const INDEX_HEAD_DIM: usize = 128;
-const CACHE_BLOCK_SIZE: usize = 64; // DeepGEMM paged MQA requires BLOCK_KV=64
-const NUM_SMS: usize = 132;
 
 fn bf16_digest(data: &[bf16]) -> String {
     let mut hasher = Sha256::new();
@@ -190,11 +189,11 @@ fn indexer_oracle_gate() -> Result<()> {
     let w = load_indexer_layer0(&ctx, &model_path)?;
 
     // ---- cache setup ----
-    let cache_blocks = ORACLE_CTX.div_ceil(CACHE_BLOCK_SIZE);
+    let cache_blocks = ORACLE_CTX.div_ceil(INDEX_CACHE_BLOCK);
     let cache_layout = Glm52IndexerCacheLayout {
         cache_blocks,
-        cache_block_size: CACHE_BLOCK_SIZE,
-        cache_block_stride_bytes: CACHE_BLOCK_SIZE * (INDEX_HEAD_DIM + 4),
+        cache_block_size: INDEX_CACHE_BLOCK,
+        cache_block_stride_bytes: INDEX_CACHE_BLOCK * (GLM52_INDEX_HEAD_DIM + 4),
     };
     let cache_bytes = cache_layout.min_cache_bytes()?;
     let mut index_k_cache = ctx.stream.alloc_zeros::<u8>(cache_bytes)?;
